requirements.txt
python-dotenv
fastapi[all]
uvicorn
pipecat-ai[openai,silero,websocket,google,daily,deepgram]>=0.0.82    
ipykernel
pipecat-ai
python-dotenv
fastapi
ngrok
pipecat-ai[groq,deepseek,google,deepgram,daily,webrtc,speechmatics]
uvicorn[standard]
ffmpeg 
twilio
pymongo
uvicorn
noisereduce
onnxruntime
websockets
pytz
motor
pipecat_ai_small_webrtc_prebuilt
aiortc
speechmatics-python
langchain_openai
motor
langchain_google_genai
langchain_mcp_adapters
langgraph
langgraph_swarm
graphiti-core
pydantic
langchain
python-dotenv
langchain_core
